+++ shipeasy/__init__.py
@@ -1,10 +1,26 @@
 from ._client import Client, ExperimentResult
+from ._detail import (
+    FlagDetail,
+    CLIENT_NOT_READY,
+    FLAG_NOT_FOUND,
+    OFF,
+    OVERRIDE,
+    RULE_MATCH,
+    DEFAULT,
+)
 from ._hash import murmur3
 from .middleware import AnonIdMiddleware, AnonIdASGIMiddleware
 
 __all__ = [
     "Client",
     "ExperimentResult",
+    "FlagDetail",
+    "CLIENT_NOT_READY",
+    "FLAG_NOT_FOUND",
+    "OFF",
+    "OVERRIDE",
+    "RULE_MATCH",
+    "DEFAULT",
     "murmur3",
     "AnonIdMiddleware",
     "AnonIdASGIMiddleware",